Change id_servico column type from BIGINT to INT

+++ scripts/v1.1__alter_table_item_atendimento.sql
@@ -1,5 +1,5 @@
 ALTER TABLE item_atendimento
-ADD COLUMN id_servico BIGINT;
+ADD COLUMN id_servico INT;
 
 ALTER TABLE item_atendimento
 ADD CONSTRAINT fk_item_servico
